[Refactor]: move generated folder to i18n folder
