+++ src/auth/auth.controller.spec.ts
@@ -7,7 +7,7 @@ import { JwtService } from "src/jwt/jwt.service";
 import { ConfigService } from "@nestjs/config";
 import { LoginDto } from "./dto/login.dto";
 import { I18nContext } from "nestjs-i18n";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 
 describe(`AuthController`, () => {
   let service: AuthService;
