+++ src/auth/auth.service.spec.ts
@@ -4,7 +4,7 @@ import { Test, TestingModule } from "@nestjs/testing";
 import { PrismaClient } from "@prisma/client";
 import * as bcrypt from "bcrypt";
 import { I18nContext } from "nestjs-i18n";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 import { JwtService } from "src/jwt/jwt.service";
 import { PrismaService } from "src/prisma/prisma.service";
 import { AuthService, PASSWORD_HASH_SALT } from "./auth.service";
