+++ src/auth/auth.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from "@nestjs/common";
 import * as bcrypt from "bcrypt";
 import { I18nContext } from "nestjs-i18n";
 import { IAccessTokenPayload } from "src/common/config/interfaces/jwt.interface";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 import { JwtService } from "src/jwt/jwt.service";
 import { PrismaService } from "src/prisma/prisma.service";
 import { LoginDto, LoginResponseDto } from "./dto/login.dto";
