+++ src/common/httpExceptions/http-exception.filter.ts
@@ -7,7 +7,7 @@ import {
 import { Response } from "express";
 import { I18nContext } from "nestjs-i18n";
 import { CoreResponseDto } from "src/common/dto/common.dto";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 
 @Catch()
 export class AllExceptionFilter implements ExceptionFilter {
