+++ src/common/utils/userVerified.ts
@@ -1,6 +1,6 @@
 import { User } from "@prisma/client";
 import { I18nContext } from "nestjs-i18n";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 import { ForbiddenException } from "@nestjs/common";
 
 export const checkIfUserIsVerified = (
