+++ src/company/company.controller.ts
@@ -11,7 +11,7 @@ import { User } from "@prisma/client";
 import { I18n, I18nContext } from "nestjs-i18n";
 import { AuthUser } from "src/common/decorators/auth-user.decorator";
 import { setRole } from "src/common/decorators/setRole.decorator";
-import { I18nTranslations } from "src/generated/i18n.generated";
+import { I18nTranslations } from "src/i18n/generated/i18n.generated";
 import { CompanyService } from "./company.service";
 import {
   CreateCompanyDto,
